+++ main.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 """MCP server entry point
 
 Registers all API tools under tools as Redmine MCP server using @mcp.tool() and starts the server.
@@ -9,9 +10,6 @@ from tools.Attachments.DeleteAttachmentTool import DeleteAttachmentTool
 from tools.Attachments.GetAttachmentTool import GetAttachmentTool
 from tools.Attachments.UpdateAttachmentTool import UpdateAttachmentTool
 from tools.Attachments.UploadAttachmentTool import UploadAttachmentTool
-from tools.get_queries_tool import GetQueriesTool
-from tools.get_versions_tool import GetVersionTool
-from tools.get_wiki_pages_tool import GetWikiPagesTool
 from tools.Issues.AddWatcherTool import AddWatcherTool
 from tools.Issues.CreateIssueTool import CreateIssueTool
 from tools.Issues.DeleteIssueTool import DeleteIssueTool
@@ -22,24 +20,27 @@ from tools.Issues.RemoveWatcherTool import RemoveWatcherTool
 from tools.Issues.UpdateIssueTool import UpdateIssueTool
 from tools.MyAccount.GetMyAccountTool import GetMyAccountTool
 from tools.News.GetNewsTool import GetNewsTool
-from tools.ProjectMemberships.create_membership_tool import CreateProjectMembershipTool
-from tools.ProjectMemberships.get_memberships_tool import GetMembershipsTool
-from tools.Projects.archive_project_tool import ArchiveProjectTool
-from tools.Projects.create_project_tool import CreateProjectTool
-from tools.Projects.delete_project_tool import DeleteProjectTool
-from tools.Projects.get_project_tool import GetProjectTool
+from tools.ProjectMemberships.CreateProjectMembershipTool import CreateProjectMembershipTool
+from tools.ProjectMemberships.GetMembershipsTool import GetMembershipsTool
+from tools.Projects.ArchiveProjectTool import ArchiveProjectTool
+from tools.Projects.CreateProjectTool import CreateProjectTool
+from tools.Projects.DeleteProjectTool import DeleteProjectTool
 from tools.Projects.get_projects_tool import GetProjectsTool
-from tools.Projects.unarchive_project_tool import UnarchiveProjectTool
-from tools.Projects.update_project_tool import UpdateProjectTool
-from tools.Roles.get_role_tool import GetRoleTool
+from tools.Projects.GetProjectTool import GetProjectTool
+from tools.Projects.UnarchiveProjectTool import UnarchiveProjectTool
+from tools.Projects.UpdateProjectTool import UpdateProjectTool
+from tools.Queries.GetQueriesTool import GetQueriesTool
 from tools.Roles.get_roles_tool import GetRolesTool
-from tools.TimeEntries.create_time_entry_tool import CreateTimeEntryTool
-from tools.TimeEntries.get_time_entries_tool import GetTimeEntriesTool
-from tools.Users.create_user_tool import CreateUserTool
-from tools.Users.delete_user_tool import DeleteUserTool
-from tools.Users.get_user_tool import GetUserTool
+from tools.Roles.GetRoleTool import GetRoleTool
+from tools.TimeEntries.CreateTimeEntryTool import CreateTimeEntryTool
+from tools.TimeEntries.GetTimeEntriesTool import GetTimeEntriesTool
+from tools.Users.CreateUserTool import CreateUserTool
+from tools.Users.DeleteUserTool import DeleteUserTool
 from tools.Users.get_users_tool import GetUsersTool
-from tools.Users.update_user_tool import UpdateUserTool
+from tools.Users.GetUserTool import GetUserTool
+from tools.Users.UpdateUserTool import UpdateUserTool
+from tools.Versions.GetVersionTool import GetVersionTool
+from tools.WikiPages.GetWikiPagesTool import GetWikiPagesTool
 
 mcp = FastMCP("Redmine MCP Server", "0.1.0")
