+++ tools/ProjectMemberships/create_membership_tool.py
@@ -13,7 +13,6 @@ Raises:
 from typing import Any, Dict, List
 
 import requests
-from fastmcp.tools.tool import Tool
 
 from tools.redmine_api_client import RedmineAPIClient
 
@@ -66,10 +65,3 @@ def create_membership(
     except Exception as e:
         # Raise other exceptions
         raise
-
-
-CreateProjectMembershipTool = Tool.from_function(
-    create_membership,
-    name="create_project_membership",
-    description="Create a new project membership in Redmine",
-)
