+++ tools/Projects/CreateProjectTool.py
@@ -1,0 +1,9 @@
+from fastmcp.tools.tool import Tool
+
+from tools.Projects.create_project_tool import create_project
+
+CreateProjectTool = Tool.from_function(
+    create_project,
+    name="create_project",
+    description="Create a Redmine project",
+)
