+++ tools/TimeEntries/CreateTimeEntryTool.py
@@ -1,0 +1,9 @@
+from fastmcp.tools.tool import Tool
+
+from tools.TimeEntries.create_time_entry_tool import create_time_entry
+
+CreateTimeEntryTool = Tool.from_function(
+    create_time_entry,
+    name="create_time_entry",
+    description="Create a new time entry in Redmine.",
+)
