+++ tools/WikiPages/GetWikiPagesTool.py
@@ -1,0 +1,9 @@
+from fastmcp.tools.tool import Tool
+
+from tools.WikiPages.get_wiki_pages_tool import get_wiki_pages
+
+GetWikiPagesTool = Tool.from_function(
+    get_wiki_pages,
+    name="get_wiki_pages",
+    description="Get wiki pages from Redmine",
+)
