+++ tools/WikiPages/get_wiki_pages_tool.py
@@ -1,19 +1,9 @@
-"""
-Query Tool
-保存済み検索条件
-"""
-
 import os
-
-from fastmcp.tools.tool import Tool
 
 from tools.redmine_api_client import RedmineAPIClient
 
 
-def get_queries(
-    redmine_url: str,
-    api_key: str,
-):
+def get_wiki_pages(redmine_url: str, api_key: str, project_id: str):
     if redmine_url is None:
         redmine_url = os.environ.get("REDMINE_URL")
     if api_key is None:
@@ -21,20 +11,16 @@ def get_queries(
     if not redmine_url or not api_key:
         raise ValueError("redmine_url and api_key are required.")
     client = RedmineAPIClient(base_url=redmine_url, api_key=api_key)
+    endpoint = f"/projects/{project_id}/wiki/index.json"
     try:
-        response = client.get(endpoint="/queries.json")
+        response = client.get(endpoint=endpoint)
         data = response.json()
         return {
-            "queries": data.get("queries", []),
+            "wiki_pages": data.get("wiki_pages", []),
         }
     except Exception as e:
+        import requests
+
         if hasattr(e, "response") and e.response is not None and getattr(e.response, "status_code", None) == 404:
-            return {"queries": []}
+            return {"wiki_pages": []}
         raise
-
-
-GetQueriesTool = Tool.from_function(
-    get_queries,
-    name="get_queries",
-    description="Get a list of queries from Redmine.",
-)
